+++ src/Sql/Sql.Test/ScenarioTests/ManagedDatabaseCrudScenarioTests.cs
@@ -92,5 +92,12 @@ namespace Microsoft.Azure.Commands.Sql.Test.ScenarioTests
         {
             RunPowerShellTest("Test-GeoRestoreManagedDatabase");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestSetManagedDatabase()
+        {
+            RunPowerShellTest("Test-SetManagedDatabase");
+        }
     }
 }